Updated Maple previewer for comma-separated content
The input experssion is now wrapped in a list to
ensure the MathML contains all content. Previously
only the first item in a sequence would be
displayed.

This method adds extra square brackets, so
these are removed. Due to changes in the way tahe
MathML is constructed between different Maple
versions, this is  does using regular expressions
to capture the correct "mfenced" tag.